pyproject.toml
[project]
name = "do-abstimmungsstudio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery[redis]>=5.6.2",
    "boto3<1.36",
    "django>=6.0.3",
    "django-celery-beat>=2.9.0",
    "django-celery-results>=2.6.0",
    "django-environ>=0.13.0",
    "django-health-check>=4.1.2",
    "django-ninja>=1.5.3",
    "django-storages[s3]>=1.14.6",
    "django-unfold>=0.83.1",
    "granian>=2.7.2",
    "influxdb-client[ciso]>=1.50.0",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "polars>=1.38.1",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=23.0.1",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "sentry-sdk>=2.54.0",
    "whitenoise>=6.12.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.5",
    "ty>=0.0.21",
]
